Improving clinical data informaiton in the vignette. Trying to solve the problem when the download stops, not all files are moved.
Former-commit-id: fe14ddc1c1d5b596a1b0939b8914b0d9e800ccbe

+++ R/TCGAInternal.R
@@ -110,10 +110,12 @@ getGDCprojects <- function(){
 
 # Source: https://stackoverflow.com/questions/10266963/moving-files-between-folders
 move <- function(from, to) {
-    todir <- dirname(to)
-    if (!isTRUE(file.info(todir)$isdir)) dir.create(todir, recursive=TRUE,showWarnings = FALSE)
-    file.rename(from = from,  to = to)
-    if(dirname(from) != ".") unlink(dirname(from),recursive=TRUE,force = TRUE)
+    if(file.exists(from)){
+        todir <- dirname(to)
+        if (!isTRUE(file.info(todir)$isdir)) dir.create(todir, recursive=TRUE,showWarnings = FALSE)
+        file.rename(from = from,  to = to)
+        if(dirname(from) != ".") unlink(dirname(from),recursive=TRUE,force = TRUE)
+    }
 }
 
 getProjectSummary <- function(project, legacy = FALSE){
